feat: 가중치 테이블 분리, 회원가입 시 가중치 테이블 초기화 이벤트 처리 방식으로 변환
