+++ build.gradle
@@ -34,7 +34,7 @@ dependencies {
 	implementation 'org.apache.commons:commons-text:1.10.0'
 //	implementation 'io.netty:netty-resolver-dns-native-macos:4.1.72.Final'
 	implementation 'io.netty:netty-resolver-dns-native-macos:4.1.68.Final:osx-aarch_64'
-	implementation 'mysql:mysql-connector-java'
+//	implementation 'mysql:mysql-connector-java'
 
 	// S3
 	implementation 'org.springframework.cloud:spring-cloud-starter-aws:2.2.6.RELEASE'
