+++ src/main/java/muse_kopis/muse/genre/domain/Genre.java
@@ -1,4 +1,4 @@
-package muse_kopis.muse.performance.domain.genre.domain;
+package muse_kopis.muse.genre.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.EnumType;
