--- src/main/java/muse_kopis/muse/performance/domain/genre/application/GenreService.java
@@ -1,8 +1,0 @@
-package muse_kopis.muse.performance.domain.genre.application;
-
-import muse_kopis.muse.performance.domain.genre.domain.GenreType;
-
-public interface GenreService {
-
-    void saveGenre(String performanceName, GenreType genreType);
-}
