+++ src/main/java/muse_kopis/muse/usergenre/application/UserGenreService.java
@@ -1,50 +1,54 @@
-package muse_kopis.muse.performance.domain.usergenre.application;
+package muse_kopis.muse.usergenre.application;
 
+import jakarta.transaction.Transactional;
 import java.util.Arrays;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import muse_kopis.muse.auth.oauth.domain.OauthMember;
 import muse_kopis.muse.auth.oauth.domain.OauthMemberRepository;
 import muse_kopis.muse.performance.domain.Performance;
 import muse_kopis.muse.performance.domain.PerformanceRepository;
 import muse_kopis.muse.performance.domain.dto.PerformanceResponse;
-import muse_kopis.muse.performance.domain.genre.domain.Genre;
-import muse_kopis.muse.performance.domain.genre.domain.GenreRepository;
-import muse_kopis.muse.performance.domain.usergenre.domain.UserGenre;
-import muse_kopis.muse.performance.domain.usergenre.domain.UserGenreRepository;
+import muse_kopis.muse.genre.domain.Genre;
+import muse_kopis.muse.genre.domain.GenreRepository;
+import muse_kopis.muse.usergenre.domain.UserGenre;
+import muse_kopis.muse.usergenre.domain.UserGenreRepository;
+import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Service;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
-public class UserGenreServiceImpl implements UserGenreService {
+public class UserGenreService {
 
     private final GenreRepository genreRepository;
     private final UserGenreRepository userGenreRepository;
     private final PerformanceRepository performanceRepository;
     private final OauthMemberRepository oauthMemberRepository;
 
-    @Override
-    public void updateGenre(Performance performance, OauthMember oauthMember) {
+    @Transactional
+    public void updateGenre(Long memberId, Long performanceId) {
+        OauthMember oauthMember = oauthMemberRepository.getByOauthMemberId(memberId);
+        Performance performance = performanceRepository.getByPerformanceId(performanceId);
         List<Genre> genre = genreRepository.findAllByPerformance(performance);
         UserGenre userGenre = userGenreRepository.findByOauthMember(oauthMember)
                 .orElseGet(() -> initGenre(oauthMember));
         genre.forEach(it -> userGenre.incrementGenreWeight(it.getGenre()));
     }
 
-    @Override
+    @Transactional
     public void updateGenres(List<Long> performanceIds, Long memberId) {
-        OauthMember oauthMember = oauthMemberRepository.getByOauthMemberId(memberId);
-        List<Performance> performances = performanceIds.stream()
-                .map(performanceRepository::getByPerformanceId).toList();
-        performances.forEach(performance -> updateGenre(performance, oauthMember));
+        performanceIds.forEach(performance -> updateGenre(performance, memberId));
     }
 
-    @Override
+    @EventListener
     public UserGenre initGenre(OauthMember oauthMember) {
+        log.info("init UserGenre table");
         return userGenreRepository.save(new UserGenre(oauthMember));
     }
 
-    @Override
+    @Transactional
     public List<PerformanceResponse> showOnboarding() {
         List<Long> performanceIds = Arrays.asList(
                 38L, 103L, 118L, 124L, 134L, 139L, 170L, 177L, 181L, 194L,
