+++ src/main/java/muse_kopis/muse/usergenre/domain/dto/UpdateInfo.java
@@ -1,0 +1,6 @@
+package muse_kopis.muse.usergenre.domain.dto;
+
+public record UpdateInfo(
+        Long performanceId
+) {
+}
